+++ input.c
@@ -115,6 +115,7 @@ static bool is_filterable_line (void);
 static bool is_filterable_field (const char* field);
 
 static bool isspaces (const char* s);
+inline bool iszero (const char* s);
 
 
 void set_input (FILE* file, char _separator, char _enclosure, bool _allow_breaks, bool _remove_bom, bool skip_after_header, size_t _skip_records, size_t _limit_records, trimmode_t _trim, filtermode_t _filter) {
@@ -487,11 +488,11 @@ inline bool is_filterable_field (const char* field) {
 		case FILTER_EMPTY:
 			return (field[0] == '\0');
 		case FILTER_ZEROES:
-			return (field[0] == '0' && field[1] == '\0');
+			return (iszero(field));
 		case FILTER_EMPTY_OR_ZEROES:
-			return (field[0] == '\0' || (field[0] == '0' && field[1] == '\0'));
+			return (field[0] == '\0' || iszero(field));
 		case FILTER_BLANK_OR_ZEROES:
-			return (field[0] == '\0' || (field[0] == '0' && field[1] == '\0') || isspaces(field));
+			return (field[0] == '\0' || iszero(field) || isspaces(field));
 		case FILTER_BLANK:
 			return (field[0] == '\0' || isspaces(field));
 		default:
@@ -505,5 +506,39 @@ inline bool isspaces (const char* s) {
 			return false;
 	}
 	return true;
+}
+
+inline bool iszero (const char* s) {
+	/* Recognized "zero" formats:
+	 *  0		S0,S1
+	 *  000		S0,S1+
+	 *  .000	S0,S2+
+	 *  0.		S0,S1,S3
+	 *  0.000	S0,S1,S3+
+	 *  000.	S0,S1+,S3
+	 *  000.000	S0,S1+,S3+
+	 */
+
+	#define NS(nextState) do{ s++; goto nextState; }while(0)
+	#define skipZeroes() do{ while(*s == '0') { s++; } }while(0)
+	#define acceptEot() do{ return (*s == '\0'); }while(0)
+
+// S0:
+	if (*s == '.')  NS(S2);
+	if (*s != '0')  return false;  // first char neither dot nor zero
+	NS(S1);
+ S1:
+	skipZeroes();
+	if (*s == '.')  NS(S3);
+	acceptEot();  // zeroes only
+ S2:
+	if (!*s)  return false;  // dot without any zeroes
+ S3:
+	skipZeroes();
+	acceptEot();
+
+ #undef NS
+ #undef skipZeroes
+ #undef acceptEot
 }